requirements.txt
asgiref==3.3.1
boto3==1.17.29
botocore==1.20.29
dj-database-url==0.5.0
Django==3.0
django-crispy==1.8.1
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
